# admin/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_admin-template_backend_internal_dto.BatchDeleteReq:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
  github_com_admin-template_backend_internal_dto.ChangeStatusReq:
    properties:
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - status
    type: object
...
  github_com_admin-template_backend_internal_dto.CreateUserReq:
    properties:
      avatar:
        maxLength: 255
        type: string
      email:
        maxLength: 128
        type: string
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      phone:
        maxLength: 32
        type: string
      remark:
        maxLength: 255
        type: string
      role_ids:
        items:
          type: integer
        type: array
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  github_com_admin-template_backend_internal_dto.ResetPasswordReq:
    properties:
      new_password:
        maxLength: 128
        minLength: 6
        type: string
    required:
    - new_password
    type: object
  github_com_admin-template_backend_internal_dto.RoleBrief:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_admin-template_backend_internal_dto.UpdateUserReq:
    properties:
      avatar:
        maxLength: 255
        type: string
      email:
        maxLength: 128
        type: string
      nickname:
        maxLength: 64
        type: string
      phone:
        maxLength: 32
        type: string
      remark:
        maxLength: 255
        type: string
      role_ids:
        items:
          type: integer
        type: array
      status:
        enum:
        - 1
        - 2
        type: integer
    type: object
  github_com_admin-template_backend_internal_dto.UserResp:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      last_login_ip:
        type: string
      nickname:
        type: string
      phone:
        type: string
      remark:
        type: string
      roles:
        items:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.RoleBrief'
        type: array
      status:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_admin-template_backend_internal_pkg_response.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_admin-template_backend_internal_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: RBAC admin backend — auth, users/roles/menus CRUD, operation log, file
    upload, Excel/CSV export.
  title: Admin Template API
  version: "1.0"
paths:
...
  /system/users:
    get:
      parameters:
      - description: 页码 (默认 1)
        in: query
        name: page
        type: integer
      - description: 每页大小 (默认 10, 最大 200)
        in: query
        name: size
        type: integer
      - description: 模糊搜索用户名/昵称/邮箱
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - System/Users
    post:
      consumes:
      - application/json
      parameters:
      - description: 新用户
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.CreateUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.UserResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
        "409":
          description: 用户名已存在
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - System/Users
  /system/users/{id}:
    delete:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 删除单个用户
      tags:
      - System/Users
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.UserResp'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 获取单个用户
      tags:
      - System/Users
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要更新的字段（指针字段，未传则不更新）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_admin-template_backend_internal_dto.UserResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - System/Users
  /system/users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: 管理员无需知道旧密码即可设置新密码。
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.ResetPasswordReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 重置用户密码
      tags:
      - System/Users
  /system/users/{id}/roles:
    post:
      consumes:
      - application/json
      description: 'body {ids: [...]} 是角色 ID 列表，传空数组清空。'
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色 ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.BatchDeleteReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 给用户分配角色
      tags:
      - System/Users
  /system/users/{id}/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 目标状态：1=启用 2=禁用
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.ChangeStatusReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 启用/禁用用户
      tags:
      - System/Users
  /system/users/batch-delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID 列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_admin-template_backend_internal_dto.BatchDeleteReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_admin-template_backend_internal_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: 批量删除用户
      tags:
      - System/Users
  /system/users/export/csv:
    get:
      parameters:
      - description: 模糊搜索（与 List 同语义）
        in: query
        name: keyword
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出用户列表为 CSV
      tags:
      - System/Users
  /system/users/export/excel:
    get:
      parameters:
      - description: 模糊搜索（与 List 同语义）
        in: query
        name: keyword
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 导出用户列表为 Excel
      tags:
      - System/Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"